+++ packages/fcl/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 - YYYY-MM-DD **BREAKING?** -- description
 
+- 2021-10-01 -- Simplify passing `currentUser` data into args with the ability to pass `currentUser` as param instead of a creating a user snapshot before.
+
+Examples of `currentUser` as a param.
+
+```javascript
+import {query, currentUser} from "@onflow/fcl"
+
+await query({
+  cadence: `
+    pub fun main(currentUser: Address): Address {
+      return currentUser
+    }
+  `,
+  args: (arg, t) => [
+    currentUser
+  ]
+})
+```
+
 ## 0.0.78-alpha.3 - 2021-10-01
 
 - 2021-10-01 -- SDK VSN 0.0.55